pyproject.toml
# pyproject.toml

[tool.poetry]
name = "neon-support-queue"
version = "0.1.0"
description = "Visualise and monitor customer support ticket wait times, with alerts for Neon DB usage thresholds."
authors = ["Lena Docherty <ang2887@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "neon_support_queue" }]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "==1.5.3"
dash = "==2.17.1"
plotly = "==5.13.1"
schedule = "==1.2.2"
python-dotenv = "==1.0.1"
Flask = "==3.0.3"
requests = "==2.32.3"
psycopg2-binary = "==2.9.9"
tenacity = "==9.0.0"
numpy = "==1.26.4"
pytz = "==2025.1"
matplotlib = "==3.10.1"
retrying = "==1.3.4"
gunicorn = "==20.1.0"
SQLAlchemy = "==1.4.39"
fonttools = "==4.56.0"
email-validator = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

